Add files via upload
I created 49 datasets with each 20 pp and 1 000 trials. These datasets are drawn from population with meanG1 = 0.3, meanG2 = 0.7 and std = 0.2 for both groups. Each dataset has thus some variation in the actual group means and stds. 
This script then plots the correlations between the power at a certain n_trial and one of the dataset-charachteristics (mean difference, pooledstd or ES). 
Goal is to visualize and gain understanding in the effect of changes is mean, ES and std on the power in the 2 power definitions.