Update Simulate_data_final.py
Working version: parameter seed still set in Generate_parameters, but now responses are generated with a seed as well (thus the response-data can be regenerated as well. For each repetition the response_seed is put to the repetition_number (see line 32). Thus possible to regenerate only 1 repetition as well for example with seed just set to the same number as the repetition number.